feat: add health command to display user defined env
