+++ src/health/mod.rs
@@ -1,6 +1,8 @@
-use std::{env, process::exit};
+use std::{env, process::{exit, Command}};
 
 use lrncore::usage_exit::command_usage;
+
+use crate::{logs, nxfs};
 
 pub fn health_help() -> String {
     let usage = r"
@@ -13,8 +15,6 @@ Options:
 
     usage.to_string()
 }
-
-
 
 pub fn health_command() {
     let args: Vec<String> = env::args().collect();
@@ -36,5 +36,20 @@ pub fn health_command() {
 }
 
 fn user_env_health() {
-    println!("user health env")
+    let config = nxfs::config::parse_config_file().expect("Failed to get the config file");
+    logs::nyx_log("User environment health status:");
+    for each in config.user.health_list {
+        match Command::new(&each)
+            .stdout(std::process::Stdio::null())
+            .stderr(std::process::Stdio::null())
+            .spawn()
+        {
+            Ok(_) => {
+                logs::installed(&each);
+            }
+            Err(_) => {
+                logs::not_installed(&each);
+            }
+        }
+    }
 }
