+++ src/logs.rs
@@ -29,3 +29,18 @@ pub fn nyx_log(msg: &str) {
     let log = "[NYX]".truecolor(138, 43, 226);
     println!("{} {}", log, msg);
 }
+
+pub fn not_installed(msg: &str) {
+    let not_installed = "[✘]".truecolor(255, 0, 0);
+    println!("\t{} {}", not_installed, msg);
+}
+
+pub fn installed(msg: &str) {
+    let installed = "[✔]".truecolor(0, 255, 0);
+    println!("\t{} {}", installed, msg);
+}
+
+pub fn warning(msg: &str) {
+    let warning = "[⚠]".truecolor(255, 155, 0);
+    println!("\t{} {}", warning, msg);
+}
